+++ src/components/FormInput/EmploymentTypeField.tsx
@@ -1,10 +1,12 @@
 /* eslint-disable @typescript-eslint/camelcase */
-import React, { FC } from 'react'
-import { Field } from 'formik'
+import React, { FC, useEffect } from 'react'
+import { Field, useFormikContext } from 'formik'
 
 import { SelectFieldProps } from '../../types'
 import { BaseSelectField } from './base'
 import { validateSelect } from './validate'
+import CompanyTax from './CompanyTaxField'
+import CompanyNameField from './CompanyNameField'
 
 const options = {
   fixed: 'umowa o pracę na czas nieokreślony',
@@ -21,28 +23,53 @@ const options = {
   service_contract: 'umowa zlecenie / umowa o dzieło',
 }
 
+const conditionValueListForCompanyNameField = [
+  'fixed',
+  'temporary',
+  'fixed_partial',
+  'government',
+  'agriculture',
+  'service_contract',
+]
+
 const EmploymentTypeField: FC<SelectFieldProps> = ({
   id,
   label,
-  name,
+  name = 'employment_type',
   required = false,
   validate,
   showError = true,
-}) => (
-  <Field
-    required={required}
-    showError={showError}
-    id={id}
-    name={name}
-    component={BaseSelectField}
-    label={label}
-    type="select"
-    options={options}
-    validate={
-      (validate && validate(required)) || validateSelect(required, options)
-    }
-  />
-)
+}) => {
+  const { values, validateForm } = useFormikContext<any>()
+  const fieldValue = values[name]
+  useEffect(() => {
+    validateForm()
+  }, [fieldValue, validateForm])
+
+  return (
+    <>
+      <Field
+        required={required}
+        showError={showError}
+        id={id}
+        name={name}
+        component={BaseSelectField}
+        label={label}
+        type="select"
+        options={options}
+        validate={
+          (validate && validate(required)) || validateSelect(required, options)
+        }
+      />
+      {name && values[name] === 'self_employed' && (
+        <CompanyTax required name="company_tax_id" />
+      )}
+      {name && conditionValueListForCompanyNameField.includes(values[name]) && (
+        <CompanyNameField required name="company_name" />
+      )}
+    </>
+  )
+}
 
 EmploymentTypeField.defaultProps = {
   name: 'employment_type',
