+++ src/components/FormInput/index.ts
@@ -19,7 +19,6 @@ export { default as EducationField } from './EducationField'
 export { default as MaritalStatusField } from './MaritalStatusField'
 export { default as HousingStatusField } from './HousingStatusField'
 export { default as EmploymentTypeField } from './EmploymentTypeField'
-export { default as EmploymentTypeField2 } from './EmploymentTypeField2'
 export { default as LoanReasonField } from './LoanReasonField'
 export { default as BirthDateField } from './BirthDateField'
 export { default as CarStatusField } from './CarStatusField'
